fix: [REV-2706] add local IP to ALLOWED_HOSTS for k8s health_check
helm-charts django-ida uses a Kubernetes readinessProbe to check whether
the container has successfully started the Django app.

The readinessProbe checks for a successful GET response from the
configured path in edx-internal's
argocd/applications/commerce-coordinator YAML config.

The pod IP must be included in ALLOWED_HOSTS for this check to succeed.

This commit uses socket.getaddrinfo to get the IPv4 and IPv6 address of
localhost and add it to ALLOWED_HOSTS.