build: add host.docker.internal to ALLOWED_HOSTS
ALLOWED_HOSTS is a Django setting for valid hostnames your app may be
aliased as.

host.docker.internal is an internal Docker alias for the host machine
running Docker Desktop. (localhost from within a Docker container is the
container itself, not the machine running Docker Desktop.)

We are currently using host.docker.internal to reach Coordinator from
devstack ecommerce.

Thus, requests arriving to localhost Coordinator may have originally
been sent to host.docker.internal:8130 instead of localhost:8130.

This configuration makes Django not reject requests originating from
Docker containers with destination to localhost.

See: https://docs.docker.com/desktop/networking/#i-want-to-connect-from-a-container-to-a-service-on-the-host