fix: remove catch of stripe.error.IdempotencyError
We needed to catch IdempotencyError before in legacy ecommerce because
we were using Stripe as our PaymentIntent source of truth instead of our
DB.

This was problematic for two reasons.

First, Stripe wipes idempotency keys 24 hours after last use. Revisiting
a payment page after 24 hours created a new PaymentIntent.

Second, Stripe would not return the PaymentIntent id if the arguments to
PaymentIntent.create() changes. For example, if a coupon is applied,
subsequent refreshes of Payment MFE would cause PaymentIntent.create()
to be called with different arguments, causing an IdempotencyError.

However, now that we are checking our order management system Titan in
Coordinator for any pre-existing PaymentIntents _before_ reaching out to
Stripe, these are no longer issues.

It is possible for there to be a race condition that generates an
IdempotencyError. A user would have to load Payment MFE twice for the
same order but with different basket prices in the backend. However,
this is much rarer, so removing this handing code until we see instances
in the wild and can determine what exactly to do when seeing this error.