Ecomm IDA Parity on Routes (#100)
* feat: Initial Payment MFE/Ecomm IDA Parity Work

This is part of THES-235.

* chore: Translating Blank First/Last Names form LMS to None

This is part of THES-237.

* chore: API Changes from Various Team Meetings

- Passing edx_lms_user_id in to create calls
- Swapping Stripe Payment Method
- referenceCode -> referenceNumber Renames

* fix: Adding CoordinatorValidationException

Adding `CoordinatorValidationException` to `CoordinatorSerializer`, this
permits the throwing of such errors so they are not caught by DRF.

* fix: Updating CaptureContext API and Serializer to unsuppressed versions

Also a minor Documentation Update

* feat: Accept HTML `Accepts:` Headers and lack of `Use-Jwt-Cookie` Header

LMS simply forwards to us, expecting HTML. We don't really 'do' HTML.
However in every case this endpoint should forward to Payment MFE. So,
unless its a 500 and uncaught thing, we will forward to the MFE.

* chore: Doc Updates, API Additions to Match specs not quite final.

- Updating Docs with Typos
- Adding edxLmsUserId to `add_item` this is pending a PR but has been
  confirmed
- Removing `complete_order` from the Create Order Call

* chore: Swapping to CoordinatorSerializer to ensure unsuppressed errors

* fix: Stripe Secret/Key Passing between Filters/Pipes

- Stripe:
  - Passing Stripe client secret for use as `key_id`
  - However, We need PaymentIntent ID for updates, as passing that too
  - Removing `secret_key_confirmation` from Update call, as this is not
    permitted
- Create order updated to return the Order UUID and not invoke
  Complete Order
- Removing extra semicolon (sorry, its been a hard knock life for me,
  ask me and ill C if i can tell you about it)

* chore: Final Removal of First/Last Name outside of Address Updates

This is part of THES-237

* chore: Test Updates

of 57, now only 7 are red.

I did have to make a change to revert a change i had made in my API
parity and i may have to validate the MFE still works.

But lets get tests 100%, then work backwards.

* chore: More Test Updates

* chore: Disabling broken test, leaving work note.

Spoke with Hammad, we didnt think this change and my massaging
of the coade forward. This is thus a "fix forward".