feat: add ValidateOrderReadyForDraftPayment
ValidateOrderReadyForDraftPayment is a new pipeline step for pipeline
traffic control.

Before this pipeline step, all steps of the pipeline were being
performed for all invocations of the filter.

This meant that, for example, an already-created payment would be
re-retrived, re-created, then updated.

ValidateOrderReadyForDraftPayment skips the relevant parts of the
pipeline that do not need processing.