feat: add middleware for logging drf exceptions
By default, Django REST Framework does not log much detail about 400 Bad
Requests that are caused by a ValidationError when a client sends a bad
payload.

Because Coordinator is an internal service and its users should (for the
most part) be non-human, it's very useful to log details about why a 400
Bad Request occured.