fix: return HTTP 204 on successful retirement calls (#454)
* fix: return HTTP 204 on successful retirement calls

The call to the user retirement endpoint doesn't actually return any
content on success. This confuses the s retirement driver, which expects a JSON
response  on a successful call that doesn't return HTTP 204 no content.

This PR modifies the API to return 204  for the user retirement driver.

* chore: updating the tests

updating the tests for the replaced status code